FUSION_HEAD/requirements.txt
torch==2.1.0
transformers==4.38.2
pandas
openpyxl
tqdm
Pillow
